requirements.txt
PySide6>=6.6
snowflake-connector-python>=3.6
